Feat/switch to real application image (#3)
* wip (local env script)

* feat(license): switch from MIT to Apache 2.0 License

The project's license has been updated from MIT to Apache 2.0 to better manage contributions and patent-related terms. This change includes updates to the LICENSE file to reflect the Apache 2.0 terms and conditions.

* feat(scripts): enhance k3s management in run script

Added checks for kubeconfig file existence and readability, set the KUBECONFIG environment variable, and introduced a new 'stop' command to manage k3s service.

* refactor(scripts): streamline install-setup.sh

Removed unnecessary checks for GITHUB_USER as an argument, updated version handling for KubeScore, and adjusted the ownership settings for the k3s.yaml file. Simplified the usage instructions and removed the option to skip ShellCheck installation to standardize the setup process.